refactor: removed hooks not use and use lp token as param query data
