+++ src/pages/Pools/hookV3.ts
@@ -12,11 +12,13 @@ import isEqual from 'lodash/isEqual';
 import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { updateLpPools } from 'reducer/token';
-import { fetchTokenInfo } from 'rest/api';
+import { fetchTokenInfo, fetchRewardPerSecInfo } from 'rest/api';
 import axios from 'rest/request';
 import { RootState } from 'store/configure';
 import { PoolInfoResponse } from 'types/pool';
 import { PoolTableData } from './';
+import { parseAssetOnlyDenom } from 'pages/Pools/helpers';
+import { RewardPoolType } from 'reducer/config';
 
 export const calculateLpPoolsV3 = (lpAddresses: string[], res: AggregateResult) => {
   const lpTokenData = Object.fromEntries(
@@ -432,4 +434,48 @@ export const useGetPoolsWithClaimableAmount = ({ poolTableData, totalRewardInfoD
   }, [totalRewardInfoData, poolTableData, cachePrices]);
 
   return listClaimable;
+};
+
+
+
+/**
+ * fetch reward asset for each pool, with unique key is staking token (also called liquidity address)
+ * @param pairs 
+ * @returns list reward. format: [{
+ *   "reward": [
+ *       "AIRI"
+ *   ],
+ *   "liquidity_token": "orai1hxm433hnwthrxneyjysvhny539s9kh6s2g2n8y"
+ * }]
+ */
+export const useFetchCacheRewardAssetForAllPools = (lpAddresses: string[]) => {
+  const [cachedReward, setCachedReward] = useConfigReducer('rewardPools');
+  const fetchReward = async () => {
+    let rewardAll: RewardPoolType[] = await Promise.all(
+      lpAddresses.map(async (lpAddress) => {
+        const rewardPerSecInfo = await fetchRewardPerSecInfo(lpAddress)
+        const reward = rewardPerSecInfo.assets.reduce((acc, rewardAsset) => {
+          const rewardDenom = parseAssetOnlyDenom(rewardAsset.info)
+          const token =
+            'token' in rewardAsset.info ? cw20TokenMap[rewardDenom] : tokenMap[rewardDenom];
+          // TODO: hardcode token reward xOCH
+          const xOCH_TOKEN_NAME = 'xOCH'
+          return [...acc, token ? token.name : xOCH_TOKEN_NAME];
+        }, []);
+        return {
+          reward,
+          liquidity_token: lpAddress
+        };
+      })
+    );
+    setCachedReward(rewardAll);
+  };
+
+  useEffect(() => {
+    if (!cachedReward || !cachedReward.length || cachedReward.length < lpAddresses?.length) {
+      fetchReward();
+    }
+  }, [lpAddresses]);
+
+  return [cachedReward];
 };
